+++ lib/compare_linker/formatter/markdown.rb
@@ -1,14 +1,9 @@
-require "ostruct"
-
 class CompareLinker
   class Formatter
     class Markdown < Base
-      attr_reader :downgraded
+      private
 
-      def format(gem_info)
-        g = OpenStruct.new(gem_info)
-        @downgraded = downgrade?(g.old_ver, g.new_ver, g.old_tag, g.new_tag, g.old_rev, g.new_rev)
-
+      def post_format
         text = "* [ ] "
         text += case
         when g.owner && g.old_rev && g.new_rev
@@ -57,22 +52,8 @@ class CompareLinker
           ]
         end
 
-        if downgraded
-          text += " (downgrade)"
-        end
-
+        text += " (downgrade)" if downgraded
         text
-      end
-
-      private
-
-      def github_url(repo_owner, repo_name)
-        "https://github.com/#{repo_owner}/#{repo_name}"
-      end
-
-      def github_compare_url(repo_owner, repo_name, old_tag, new_tag)
-        range = downgraded ? "#{new_tag}...#{old_tag}" : "#{old_tag}...#{new_tag}"
-        "#{github_url(repo_owner, repo_name)}/compare/#{range}"
       end
     end
   end
