+++ lib/compare_linker/formatter/text.rb
@@ -1,14 +1,9 @@
-require "ostruct"
-
 class CompareLinker
   class Formatter
     class Text < Base
-      attr_reader :downgraded
+      private
 
-      def format(gem_info)
-        g = OpenStruct.new(gem_info)
-        @downgraded = downgrade?(g.old_ver, g.new_ver, g.old_tag, g.new_tag, g.old_rev, g.new_rev)
-
+      def post_format
         text = case
         when g.owner && g.old_rev && g.new_rev
           "#{g.gem_name}: #{github_compare_url(g.owner, g.gem_name, g.old_rev, g.new_rev)}"
@@ -24,22 +19,8 @@ class CompareLinker
           "#{g.gem_name} (link not found): #{g.old_ver}...#{g.new_ver}"
         end
 
-        if downgraded
-          text += " (downgrade)"
-        end
-
+        text += " (downgrade)" if downgraded
         text
-      end
-
-      private
-
-      def github_url(repo_owner, repo_name)
-        "https://github.com/#{repo_owner}/#{repo_name}"
-      end
-
-      def github_compare_url(repo_owner, repo_name, old_tag, new_tag)
-        range = downgraded ? "#{new_tag}...#{old_tag}" : "#{old_tag}...#{new_tag}"
-        "#{github_url(repo_owner, repo_name)}/compare/#{range}"
       end
     end
   end
